requirements.txt
charset-normalizer==3.3.2
colorama==0.4.6
filelock==3.9.0
fsspec==2023.12.2
huggingface-hub==0.20.2
idna==3.6
Jinja2==3.1.2
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.0
#numpy==1.26.3
packaging==23.2
pillow==10.2.0
protobuf==4.25.2
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
safetensors==0.4.1
sentencepiece==0.1.99
sympy==1.12
tokenizers==0.15.0
#torch==2.1.2+cu118
tqdm==4.66.1
transformers==4.36.2
typing_extensions==4.4.0
urllib3==2.1.0
jsonlines